double oopsies!!! since the .css file is in the same folder as those image, I didn't need to add `assets` before their file names
I will be doing the same for any file link in the js file in the future, if I have one. This is definitely why the thing didn't work for github-actions. but how is it that npm preview and liveserver were able to load the images  still? anyways vite has a long way to go in terms of this local-file-source-string of a thing. also notice how this didn't affect the images linked directly in the html.